+++ core/src/test/kotlin/net/corda/core/flows/ReceiveFinalityFlowTest.kt
@@ -8,7 +8,7 @@ import net.corda.core.utilities.OpaqueBytes
 import net.corda.core.utilities.getOrThrow
 import net.corda.finance.GBP
 import net.corda.finance.POUNDS
-import net.corda.finance.contracts.getCashBalance
+import net.corda.finance.workflows.getCashBalance
 import net.corda.finance.flows.CashIssueAndPaymentFlow
 import net.corda.finance.flows.CashPaymentReceiverFlow
 import net.corda.node.services.statemachine.StaffedFlowHospital.*