+++ samples/irs-demo/src/integration-test/kotlin/net/corda/irs/IRSDemoTest.kt
@@ -19,7 +19,7 @@ import net.corda.core.utilities.NetworkHostAndPort
 import net.corda.core.utilities.contextLogger
 import net.corda.core.utilities.getOrThrow
 import net.corda.core.utilities.seconds
-import net.corda.finance.plugin.registerFinanceJSONMappers
+import net.corda.finance.workflows.plugin.registerFinanceJSONMappers
 import net.corda.irs.contract.InterestRateSwap
 import net.corda.irs.web.IrsDemoWebApplication
 import net.corda.test.spring.springDriver
@@ -34,6 +34,8 @@ import net.corda.testing.internal.IntegrationTestSchemas
 import net.corda.testing.internal.toDatabaseSchemaName
 import net.corda.testing.node.NotarySpec
 import net.corda.testing.node.User
+import net.corda.testing.node.internal.FINANCE_CORDAPPS
+import net.corda.testing.node.internal.cordappWithPackages
 import org.apache.commons.io.IOUtils
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.ClassRule
@@ -66,13 +68,13 @@ class IRSDemoTest : IntegrationTest() {
         springDriver(DriverParameters(
                 useTestClock = true,
                 notarySpecs = listOf(NotarySpec(DUMMY_NOTARY_NAME, rpcUsers = rpcUsers)),
-                extraCordappPackagesToScan = listOf("net.corda.irs", "net.corda.finance", "migration")
+                cordappsForAllNodes = FINANCE_CORDAPPS + cordappWithPackages("net.corda.irs")
         )) {
-            val (notary, nodeA, nodeB, controller) = listOf(
-                    defaultNotaryNode,
+            val (nodeA, nodeB, controller) = listOf(
                     startNode(providedName = DUMMY_BANK_A_NAME, rpcUsers = rpcUsers),
                     startNode(providedName = DUMMY_BANK_B_NAME, rpcUsers = rpcUsers),
-                    startNode(providedName = CordaX500Name("Regulator", "Moscow", "RU"))
+                    startNode(providedName = CordaX500Name("Regulator", "Moscow", "RU")),
+                    defaultNotaryNode
             ).map { it.getOrThrow() }
 
             log.info("All nodes started")